+++ silence_fixit_error.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import argparse
+import re
 import subprocess
 import sys
 from collections import defaultdict
@@ -9,9 +10,25 @@ from collections.abc import Sequence
 from typing import NamedTuple
 
 
+ERROR_LINE_RE = re.compile(r'^.*?@\d+:\d+ ')
+
+
 class Violation(NamedTuple):
+    filename: str
     rule_name: str
     lineno: int
+
+
+def _parse_output_line(line: str) -> Violation | None:
+    if not ERROR_LINE_RE.match(line):
+        return None
+
+    location, violated_rule_name, *__ = line.split(maxsplit=2)
+    filename, position = location.split('@', maxsplit=1)
+    lineno, *__ = position.split(':', maxsplit=1)
+
+    rule_name_ = violated_rule_name.removesuffix(':')
+    return Violation(filename, rule_name_, int(lineno))
 
 
 def _find_violations(
@@ -30,13 +47,11 @@ def _find_violations(
     # extract filenames and line numbers
     results: dict[str, list[Violation]] = defaultdict(list)
     for line in proc.stdout.splitlines():
-        location, violated_rule_name, __ = line.split(maxsplit=2)
-        filename, position = location.split('@', maxsplit=1)
-        lineno, __ = position.split(':', maxsplit=1)
-
-        rule_name_ = violated_rule_name.removesuffix(':')
-
-        results[filename].append(Violation(rule_name_, int(lineno)))
+        violation = _parse_output_line(line)
+        if violation:
+            results[violation.filename].append(violation)
+        else:  # pragma: no cover
+            pass
 
     return results